[Terminal] separate out concerns
- separate out concerns
- create headers fror utils and folder
